feat(everything): full rewrite of functionality
Rewrote entire functionality. Now follows route -> logic -> data for functionality, with a handler
file for each. DB and redis access are handled together so all data handling is abstracted away from
the rest of the app. Routing logic assumes it gets "whatever it expects to return" or an error via a
callback to logic, and logic handles anything to be done between routing and data. Sometimes there
isn't much to this, but its a nice division in either case. Added redis and versioning. More utils
including a typeCheck to ensure that when type is boolean, all values passed in are indeed JS
booleans.

BREAKING CHANGE:
Full rewrite